+++ ios/Classes/InAppWebView/InAppWebView.swift
@@ -2042,9 +2042,13 @@ public class InAppWebView: WKWebView, UIScrollViewDelegate, WKUIDelegate,
                 return
             }
             
-            if let scheme = challenge.protectionSpace.protocol, scheme == "https",
-               let sslCertificate = challenge.protectionSpace.sslCertificate {
-                InAppWebView.sslCertificatesMap[challenge.protectionSpace.host] = sslCertificate
+            // workaround for ProtectionSpace SSL Certificate
+            // https://github.com/pichillilorenzo/flutter_inappwebview/issues/1678
+            DispatchQueue.global(qos: .background).async {
+                if let scheme = challenge.protectionSpace.protocol, scheme == "https",
+                   let sslCertificate = challenge.protectionSpace.sslCertificate {
+                    InAppWebView.sslCertificatesMap[challenge.protectionSpace.host] = sslCertificate
+                }
             }
             
             let callback = WebViewChannelDelegate.ReceivedServerTrustAuthRequestCallback()