+++ ios/Classes/InAppWebView/WebViewChannelDelegate.swift
@@ -886,7 +886,7 @@ public class WebViewChannelDelegate : ChannelDelegate {
             callback.defaultBehaviour(nil)
             return
         }
-        channel.invokeMethod("onPermissionRequest", arguments: request.toMap(), callback: callback);
+        channel.invokeMethod("onPermissionRequest", arguments: request.toMap(), callback: callback)
     }
     
     public class ShouldOverrideUrlLoadingCallback : BaseCallbackResult<WKNavigationActionPolicy> {
@@ -910,7 +910,7 @@ public class WebViewChannelDelegate : ChannelDelegate {
             callback.defaultBehaviour(nil)
             return
         }
-        channel.invokeMethod("shouldOverrideUrlLoading", arguments: navigationAction.toMap(), callback: callback);
+        channel.invokeMethod("shouldOverrideUrlLoading", arguments: navigationAction.toMap(), callback: callback)
     }
     
     public func onLoadStart(url: String?) {
@@ -965,7 +965,14 @@ public class WebViewChannelDelegate : ChannelDelegate {
             callback.defaultBehaviour(nil)
             return
         }
-        channel.invokeMethod("onReceivedHttpAuthRequest", arguments: challenge.toMap(), callback: callback)
+        // workaround for ProtectionSpace.toMap() SSL Certificate
+        // https://github.com/pichillilorenzo/flutter_inappwebview/issues/1678
+        DispatchQueue.global(qos: .background).async {
+            let arguments = challenge.toMap()
+            DispatchQueue.main.async {
+                channel.invokeMethod("onReceivedHttpAuthRequest", arguments: arguments, callback: callback)
+            }
+        }
     }
     
     public class ReceivedServerTrustAuthRequestCallback : BaseCallbackResult<ServerTrustAuthResponse> {
@@ -986,7 +993,14 @@ public class WebViewChannelDelegate : ChannelDelegate {
             callback.defaultBehaviour(nil)
             return
         }
-        channel.invokeMethod("onReceivedServerTrustAuthRequest", arguments: challenge.toMap(), callback: callback);
+        // workaround for ProtectionSpace.toMap() SSL Certificate
+        // https://github.com/pichillilorenzo/flutter_inappwebview/issues/1678
+        DispatchQueue.global(qos: .background).async {
+            let arguments = challenge.toMap()
+            DispatchQueue.main.async {
+                channel.invokeMethod("onReceivedServerTrustAuthRequest", arguments: arguments, callback: callback)
+            }
+        }
     }
     
     public class ReceivedClientCertRequestCallback : BaseCallbackResult<ClientCertResponse> {
@@ -1007,7 +1021,14 @@ public class WebViewChannelDelegate : ChannelDelegate {
             callback.defaultBehaviour(nil)
             return
         }
-        channel.invokeMethod("onReceivedClientCertRequest", arguments: challenge.toMap(), callback: callback);
+        // workaround for ProtectionSpace.toMap() SSL Certificate
+        // https://github.com/pichillilorenzo/flutter_inappwebview/issues/1678
+        DispatchQueue.global(qos: .background).async {
+            let arguments = challenge.toMap()
+            DispatchQueue.main.async {
+                channel.invokeMethod("onReceivedClientCertRequest", arguments: arguments, callback: callback)
+            }
+        }
     }
     
     public func onZoomScaleChanged(newScale: Float, oldScale: Float) {
@@ -1061,7 +1082,7 @@ public class WebViewChannelDelegate : ChannelDelegate {
             "handlerName": handlerName,
             "args": args
         ]
-        channel.invokeMethod("onCallJsHandler", arguments: arguments, callback: callback);
+        channel.invokeMethod("onCallJsHandler", arguments: arguments, callback: callback)
     }
     
     public class NavigationResponseCallback : BaseCallbackResult<WKNavigationResponsePolicy> {
@@ -1085,7 +1106,7 @@ public class WebViewChannelDelegate : ChannelDelegate {
             callback.defaultBehaviour(nil)
             return
         }
-        channel.invokeMethod("onNavigationResponse", arguments: navigationResponse.toMap(), callback: callback);
+        channel.invokeMethod("onNavigationResponse", arguments: navigationResponse.toMap(), callback: callback)
     }
     
     public class ShouldAllowDeprecatedTLSCallback : BaseCallbackResult<Bool> {
@@ -1109,7 +1130,14 @@ public class WebViewChannelDelegate : ChannelDelegate {
             callback.defaultBehaviour(nil)
             return
         }
-        channel.invokeMethod("shouldAllowDeprecatedTLS", arguments: challenge.toMap(), callback: callback)
+        // workaround for ProtectionSpace.toMap() SSL Certificate
+        // https://github.com/pichillilorenzo/flutter_inappwebview/issues/1678
+        DispatchQueue.global(qos: .background).async {
+            let arguments = challenge.toMap()
+            DispatchQueue.main.async {
+                channel.invokeMethod("shouldAllowDeprecatedTLS", arguments: arguments, callback: callback)
+            }
+        }
     }
     
     public func onWebContentProcessDidTerminate() {